requirements.txt
google-genai
python-dotenv
pandas
z3-solver
flask
gunicorn
scikit-learn
matplotlib